requirements.txt
lark-parser
trimesh
numpy
pyglet<2
networkx
scipy